+++ Sources/KeyboardKit/App/KeyboardApp.swift
@@ -25,7 +25,7 @@ import SwiftUI
 ///             bundleId: "com.keyboardkit.demo",
 ///             appGroupId: "group.com.keyboardkit.demo",
 ///             locales: [.english, .swedish, .persian],
-///             dictationDeepLink: "keyboardkit://dictation"
+///             deepLinks: .init(app: "keyboardkit://")
 ///         )
 ///     }
 /// }
@@ -46,40 +46,103 @@ public struct KeyboardApp {
     ///   - name: The name of the app.
     ///   - licenseKey: Your license key, if any.
     ///   - bundleId: The app's bundle identifier.
+    ///   - keyboardExtensionBundleId: The app's keyboard bundle identifier, by default the bundle ID with a `.keyboard` suffix.
     ///   - appGroupId: The app's App Group identifier, if any.
     ///   - locales: The locales to use in the app, by default `.all`.
-    ///   - dictationDeepLink: The app's dictation deep link, if any.
+    ///   - deepLinks: App-specific deep links, if any.
     public init(
         name: String,
-        licenseKey: String? = "",
+        licenseKey: String = "",
         bundleId: String,
-        appGroupId: String? = "",
+        keyboardExtensionBundleId: String? = nil,
+        appGroupId: String? = nil,
         locales: [KeyboardLocale] = .all,
-        dictationDeepLink: String? = ""
+        deepLinks: DeepLinks? = nil
     ) {
         self.name = name
         self.bundleId = bundleId
+        self.keyboardExtensionBundleId = keyboardExtensionBundleId ?? "\(bundleId).keyboard"
         self.appGroupId = appGroupId
         self.locales = locales
         self.licenseKey = licenseKey
-        if let appGroupId, let dictationDeepLink {
+        self.deepLinks = deepLinks
+        if let appGroupId, let dictationLink = deepLinks?.dictation {
             dictationConfiguration = .init(
                 appGroupId: appGroupId,
-                appDeepLink: dictationDeepLink
+                appDeepLink: dictationLink
             )
         } else {
             dictationConfiguration = nil
         }
     }
 
+    @available(*, deprecated, message: "Use the deepLinks initializer instead.")
+    public init(
+        name: String,
+        licenseKey: String = "",
+        bundleId: String,
+        keyboardExtensionBundleId: String? = nil,
+        appGroupId: String? = nil,
+        locales: [KeyboardLocale] = .all,
+        dictationDeepLink: String? = ""
+    ) {
+        self.init(
+            name: name,
+            licenseKey: licenseKey,
+            bundleId: bundleId,
+            keyboardExtensionBundleId: keyboardExtensionBundleId,
+            appGroupId: appGroupId,
+            locales: locales,
+            deepLinks: .init(app: "", dictation: dictationDeepLink)
+        )
+    }
+
+    /// This type can define app-specific deep links.
+    public struct DeepLinks {
+
+        /// Create a custom keyboard deep links value.
+        ///
+        /// If you do not provide a value for a certain link,
+        /// a default link will be used to guide your design.
+        ///
+        /// - Parameters:
+        ///   - app: A deep link for opening the app, e.g. `x://`.
+        ///   - dictation: A deep link for opening the app and starting dictation, by default `x://dictation`.
+        ///   - keyboardSettings: A deep link for opening the app's keyboard settings screen, by default `x://keyboardSettings`.
+        ///   - languageSettings: A deep link for opening the app's language settings screen, by default `x://languageSettings`.
+        ///   - themeSettings: A deep link for opening the app's theme settings screen, by default `x://themeSettings`.
+        public init(
+            app: String,
+            dictation: String? = nil,
+            keyboardSettings: String? = nil,
+            languageSettings: String? = nil,
+            themeSettings: String? = nil
+        ) {
+            self.app = app
+            self.dictation = dictation ?? "\(app)/dictation"
+            self.keyboardSettings = keyboardSettings ?? "\(app)/keyboardSettings"
+            self.languageSettings = languageSettings ?? "\(app)/languageSettings"
+            self.themeSettings = themeSettings ?? "\(app)/themeSettings"
+        }
+
+        public let app: String
+        public let dictation: String
+        public let keyboardSettings: String
+        public let languageSettings: String
+        public let themeSettings: String
+    }
+
     /// The name of the app.
     public let name: String
 
     /// Your license key, if any.
-    public let licenseKey: String?
+    public let licenseKey: String
 
     /// The app's bundle identifier.
     public let bundleId: String
+
+    /// The app's bundle identifier.
+    public let keyboardExtensionBundleId: String
 
     /// The app's App Group identifier, if any.
     public let appGroupId: String?
@@ -87,8 +150,20 @@ public struct KeyboardApp {
     /// The locales to use in the app.
     public let locales: [KeyboardLocale]
 
-    /// The app's dictation deep link, if any.
+    /// App-specific deep links, if any.
+    public let deepLinks: DeepLinks?
+
+    /// The app's dictation configuration, if any.
     public let dictationConfiguration: Dictation.KeyboardConfiguration?
+}
+
+public extension KeyboardApp {
+
+    /// The keyboard extension bundle ID wildcard, which can
+    /// be used to see if the keyboard extension is enabled.
+    var keyboardExtensionBundleIdWildcard: String {
+        "\(bundleId).*"
+    }
 }
 
 private extension KeyboardApp {
@@ -98,9 +173,10 @@ private extension KeyboardApp {
             name: "KeyboardKit",
             licenseKey: "abc123",
             bundleId: "com.keyboardkit.demo",
+            keyboardExtensionBundleId: "com.keyboardkit.demo.keyboard",
             appGroupId: "group.com.keyboardkit.demo",
             locales: [.english, .swedish, .persian],
-            dictationDeepLink: "keyboardkit://dictation"
+            deepLinks: .init(app: "keyboardkit://")
         )
     }
 }