+++ Sources/KeyboardKit/_Essentials/Views/Keyboard+Button.swift
@@ -18,8 +18,9 @@ public extension Keyboard {
     /// You can use an optional `contentConfig` view builder
     /// to customize or replace the button content view.
     ///
-    /// You can turn any view into a button, by applying the
-    /// `.keyboardButton(...)` view modifier.
+    /// > Information: You'll be able to style the view with
+    /// by applying a ``Keyboard/ButtonStyle`` modifier, but
+    /// this won't be implemented until KeyboardKit 9.0.
     struct Button<Content: View>: View {
         
         /// Create a keyboard button.
@@ -27,7 +28,7 @@ public extension Keyboard {
         /// - Parameters:
         ///   - action: The keyboard action to apply.
         ///   - actionHandler: The action handler to use.
-        ///   - styleProvider: The style provider to use.
+        ///   - styleService: The style service to use.
         ///   - keyboardContext: The keyboard context to which the button should apply.
         ///   - calloutContext: The callout context to affect, if any.
         ///   - edgeInsets: The edge insets to apply to the interactable area, if any.
@@ -37,7 +38,7 @@ public extension Keyboard {
         public init(
             action: KeyboardAction,
             actionHandler: KeyboardActionHandler,
-            styleProvider: KeyboardStyleProvider,
+            styleService: KeyboardStyleService,
             keyboardContext: KeyboardContext,
             calloutContext: CalloutContext?,
             edgeInsets: EdgeInsets = .init(),
@@ -47,7 +48,7 @@ public extension Keyboard {
         ) {
             self.action = action
             self.actionHandler = actionHandler
-            self.styleProvider = styleProvider
+            self.styleService = styleService
             self.keyboardContext = keyboardContext
             self.calloutContext = calloutContext
             self.edgeInsets = edgeInsets
@@ -61,7 +62,7 @@ public extension Keyboard {
         /// - Parameters:
         ///   - action: The keyboard action to apply.
         ///   - actionHandler: The action handler to use.
-        ///   - styleProvider: The style provider to use.
+        ///   - styleService: The style service to use.
         ///   - keyboardContext: The keyboard context to which the button should apply.
         ///   - calloutContext: The callout context to affect, if any.
         ///   - edgeInsets: The edge insets to apply to the interactable area, if any.
@@ -70,7 +71,7 @@ public extension Keyboard {
         public init(
             action: KeyboardAction,
             actionHandler: KeyboardActionHandler,
-            styleProvider: KeyboardStyleProvider,
+            styleService: KeyboardStyleService,
             keyboardContext: KeyboardContext,
             calloutContext: CalloutContext?,
             edgeInsets: EdgeInsets = .init(),
@@ -80,7 +81,7 @@ public extension Keyboard {
             self.init(
                 action: action,
                 actionHandler: actionHandler,
-                styleProvider: styleProvider,
+                styleService: styleService,
                 keyboardContext: keyboardContext,
                 calloutContext: calloutContext,
                 edgeInsets: edgeInsets,
@@ -92,7 +93,7 @@ public extension Keyboard {
         
         private let action: KeyboardAction
         private let actionHandler: KeyboardActionHandler
-        private let styleProvider: KeyboardStyleProvider
+        private let styleService: KeyboardStyleService
         private let keyboardContext: KeyboardContext
         private let calloutContext: CalloutContext?
         private let edgeInsets: EdgeInsets
@@ -118,6 +119,52 @@ public extension Keyboard {
                     repeatTimer: repeatTimer
                 )
         }
+
+
+        // MARK: - Deprecated
+
+        @available(*, deprecated, message: "Use the styleService initializer instead.")
+        public init(
+            action: KeyboardAction,
+            actionHandler: KeyboardActionHandler,
+            styleProvider: KeyboardStyleService,
+            keyboardContext: KeyboardContext,
+            calloutContext: CalloutContext?,
+            edgeInsets: EdgeInsets = .init(),
+            isPressed: Binding<Bool>? = nil,
+            @ViewBuilder content: @escaping ContentBuilder
+        ) {
+            self.action = action
+            self.actionHandler = actionHandler
+            self.styleService = styleProvider
+            self.keyboardContext = keyboardContext
+            self.calloutContext = calloutContext
+            self.edgeInsets = edgeInsets
+            self.isPressed = isPressed
+            self.content = content
+        }
+
+        @available(*, deprecated, message: "Use the styleService initializer instead.")
+        public init(
+            action: KeyboardAction,
+            actionHandler: KeyboardActionHandler,
+            styleProvider: KeyboardStyleService,
+            keyboardContext: KeyboardContext,
+            calloutContext: CalloutContext?,
+            edgeInsets: EdgeInsets = .init(),
+            isPressed: Binding<Bool>? = nil
+        ) where Content == Keyboard.ButtonContent {
+            self.init(
+                action: action,
+                actionHandler: actionHandler,
+                styleService: styleProvider,
+                keyboardContext: keyboardContext,
+                calloutContext: calloutContext,
+                edgeInsets: edgeInsets,
+                isPressed: isPressed,
+                content: { $0 }
+            )
+        }
     }
 }
 
@@ -127,14 +174,14 @@ private extension Keyboard.Button {
         content(
             Keyboard.ButtonContent(
                 action: action,
-                styleProvider: styleProvider,
+                styleService: styleService,
                 keyboardContext: keyboardContext
             )
         )
     }
     
     var style: Keyboard.ButtonStyle {
-        styleProvider.buttonStyle(
+        styleService.buttonStyle(
             for: action,
             isPressed: isPressed?.wrappedValue ?? isPressedInternal
         )
@@ -152,7 +199,7 @@ private extension Keyboard.Button {
             Keyboard.Button(
                 action: action,
                 actionHandler: .preview,
-                styleProvider: .preview,
+                styleService: .preview,
                 keyboardContext: .preview,
                 calloutContext: .preview
             ) {
@@ -171,7 +218,7 @@ private extension Keyboard.Button {
                 Keyboard.Button(
                     action: .emoji(.init("😀")),
                     actionHandler: .preview,
-                    styleProvider: .preview,
+                    styleService: .preview,
                     keyboardContext: .preview,
                     calloutContext: .preview,
                     edgeInsets: .init(top: 10, leading: 20, bottom: 30, trailing: 0),